move login classes to right directory
